moved wmts and ais libraries into folder
