+++ Assets/Scripts/AIS/AISShipTypes.js
