+++ Assets/Scripts/WMTS/WMTSCustomDefinitions.js
