+++ Assets/Scripts/WMTS/WMTSTileManager.js
